feat: add webpack
