+++ nodemon.json
@@ -1,6 +1,6 @@
 {
   "watch": ["src"],
   "ext": ".ts,.js",
-  "ignore": [],
+  "ignore": ["src/__tests__"],
   "exec": "node --no-warnings=ExperimentalWarning --loader ts-node/esm src/index.ts"
 }
