+++ webpack.config.cjs
@@ -1,0 +1,26 @@
+const path = require('path');
+const { CleanWebpackPlugin } = require('clean-webpack-plugin');
+
+module.exports = {
+  context: path.resolve(__dirname, 'src'),
+  entry: './index.ts',
+  output: {
+    filename: 'index.cjs',
+    path: path.resolve(__dirname, 'dist'),
+  },
+  module: {
+    rules: [
+      {
+        test: /\.tsx?$/,
+        use: 'ts-loader',
+        exclude: /node_modules/,
+      },
+    ],
+  },
+  resolve: {
+    extensions: ['.tsx', '.ts', '.js'],
+    extensionAlias: { '.js': ['.ts', '.js'] },
+  },
+  target: 'node',
+  plugins: [new CleanWebpackPlugin()],
+};
